backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
SQLAlchemy==2.0.23
transformers==4.35.2
torch==2.1.1
torchaudio==2.1.1
python-docx==1.1.0
httpx==0.25.1
pydantic==2.5.0
pydantic-settings==2.1.0
